test: isolate loopback callbacks to e2e builds
