+++ playwright.config.ts
@@ -16,7 +16,7 @@ export default defineConfig({
   },
   webServer: [
     {
-      command: `rm -rf data/e2e && pnpm build && go run ./apps/api/cmd/clickclack serve --addr 127.0.0.1:${e2ePort} --data ./data/e2e --dev-bootstrap=true --embed-frame-ancestors ${embedHostOrigin}`,
+      command: `rm -rf data/e2e && pnpm build && go run -tags clickclack_e2e_unsafe_callbacks ./apps/api/cmd/clickclack serve --addr 127.0.0.1:${e2ePort} --data ./data/e2e --dev-bootstrap=true --embed-frame-ancestors ${embedHostOrigin}`,
       url: `http://127.0.0.1:${e2ePort}`,
       reuseExistingServer: process.env.CLICKCLACK_REUSE_E2E_SERVER === "1",
       // A cold production SPA build can exceed two minutes before the Go server starts.
